+++ modules/bungeecord/src/main/java/net/analyse/bungeecord/AnalyseBungee.java
@@ -59,10 +59,6 @@ public class AnalyseBungee extends Plugin implements AnalysePlatform {
       return;
     }
 
-    if (pluginConfig.hasInvalidSendMode()) {
-      getLogger().warning("Invalid sendMode in config.json, falling back to SINGLE");
-    }
-
     // Initialize session manager
     sessionManager = new SessionManager();